knuth-bendix: use Gilman digraph
