+++ include/libsemigroups/digraph-helper.hpp
@@ -100,9 +100,15 @@ namespace libsemigroups {
       template <typename T>
       using stack_type  = std::stack<std::pair<node_type<T>, label_type<T>>>;
       using lookup_type = std::vector<uint8_t>;
+      template <typename T>
+      using topological_sort_type = std::vector<node_type<T>>;
 
       // Helper function for the two versions of is_acyclic below.
       // Not noexcept because std::stack::emplace isn't
+      // This function does not really need to exist any longer, since
+      // topological_sort can be used for the same computation, but we retain
+      // it because it was already written and uses less space than
+      // topological_sort.
       template <typename T>
       bool is_acyclic(ActionDigraph<T> const& ad,
                       stack_type<T>&          stck,
@@ -113,9 +119,10 @@ namespace libsemigroups {
           if (seen[p.first] == 2) {
             return false;
           } else if (seen[p.first] == 1 || p.second >= M) {
+            // backtrack
             seen[p.first] = 1;
             stck.pop();
-            seen[stck.top().first] = 0;
+            seen[stck.top().first] = 3;
             stck.top().second++;
           } else {
             node_type<T> node;
@@ -127,7 +134,58 @@ namespace libsemigroups {
             }
           }
         } while (stck.size() > 1);
+        seen[stck.top().first] = 1;
         return true;
+      }
+
+      // helper function for the two public methods below
+      template <typename T>
+      bool topological_sort(ActionDigraph<T> const&   ad,
+                            stack_type<T>&            stck,
+                            lookup_type&              seen,
+                            topological_sort_type<T>& order) {
+        node_type<T>  m;
+        node_type<T>  n;
+        label_type<T> e;
+      dive:
+        LIBSEMIGROUPS_ASSERT(!stck.empty());
+        LIBSEMIGROUPS_ASSERT(seen[stck.top().first] == 0);
+        m       = stck.top().first;
+        seen[m] = 2;
+        e       = 0;
+        do {
+        rise:
+          std::tie(n, e) = ad.unsafe_next_neighbor(m, e);
+          if (n != UNDEFINED) {
+            if (seen[n] == 0) {
+              // never saw this node before, so dive
+              stck.emplace(n, 0);
+              goto dive;
+            } else if (seen[n] == 1) {
+              // => all descendants of n prev. explored and no cycles found
+              // => try the next neighbour of m.
+              ++e;
+            } else {
+              LIBSEMIGROUPS_ASSERT(seen[n] == 2);
+              // => n is an ancestor and a descendant of m
+              // => there's a cycle
+              order.clear();
+              return false;
+            }
+          }
+        } while (e < ad.out_degree());
+        // => all descendants of m were explored, and no cycles were found
+        // => backtrack
+        seen[m] = 1;
+        order.push_back(m);
+        stck.pop();
+        if (stck.size() == 0) {
+          return true;
+        } else {
+          m = stck.top().first;
+          e = stck.top().second;
+          goto rise;
+        }
       }
     }  // namespace detail
 
@@ -167,15 +225,57 @@ namespace libsemigroups {
       node_type<T> const                                 N = ad.nr_nodes();
       std::stack<std::pair<node_type<T>, label_type<T>>> stck;
       std::vector<uint8_t>                               seen(N, 0);
+
       for (node_type<T> m = 0; m < N; ++m) {
         if (seen[m] == 0) {
           stck.emplace(m, 0);
           if (!detail::is_acyclic(ad, stck, seen)) {
+            LIBSEMIGROUPS_ASSERT(seen[m] != 0);
             return false;
           }
+          LIBSEMIGROUPS_ASSERT(seen[m] != 0);
         }
       }
       return true;
+    }
+
+    template <typename T>
+    detail::topological_sort_type<T>
+    topological_sort(ActionDigraph<T> const& ad) {
+      using node_type = typename ActionDigraph<T>::node_type;
+
+      size_t const                     N = ad.nr_nodes();
+      detail::stack_type<T>            stck;
+      std::vector<uint8_t>             seen(N, 0);
+      detail::topological_sort_type<T> order;
+
+      for (node_type m = 0; m < N; ++m) {
+        if (seen[m] == 0) {
+          stck.emplace(m, 0);
+          if (!detail::topological_sort(ad, stck, seen, order)) {
+            // digraph is not acyclic and so there's no topological order for
+            // the nodes.
+            LIBSEMIGROUPS_ASSERT(order.empty());
+            return order;
+          }
+        }
+      }
+      LIBSEMIGROUPS_ASSERT(order.size() == ad.nr_nodes());
+      return order;
+    }
+
+    // Topologically sort those nodes reachable from source
+    template <typename T>
+    detail::topological_sort_type<T>
+    topological_sort(ActionDigraph<T> const& ad, node_type<T> const source) {
+      size_t const                     N = ad.nr_nodes();
+      detail::stack_type<T>            stck;
+      std::vector<uint8_t>             seen(N, 0);
+      detail::topological_sort_type<T> order;
+
+      stck.emplace(source, 0);
+      detail::topological_sort(ad, stck, seen, order);
+      return order;
     }
 
     //! Check if the subdigraph induced by the nodes reachable from a source
