+++ src/knuth-bendix.cpp
@@ -20,6 +20,7 @@
 #include <string>   // for string
 
 #include "libsemigroups/cong-intf.hpp"  // for CongruenceInterface, CongruenceInterface::...
+#include "libsemigroups/digraph.hpp"              // for ActionDigraph
 #include "libsemigroups/fpsemi-intf.hpp"          // for FpSemigroupInterface
 #include "libsemigroups/froidure-pin-base.hpp"    // for FroidurePinBase
 #include "libsemigroups/froidure-pin.hpp"         // for FroidurePin
@@ -41,6 +42,27 @@ namespace libsemigroups {
   using froidure_pin_type
       = FroidurePin<detail::KBE,
                     FroidurePinTraits<detail::KBE, fpsemigroup::KnuthBendix>>;
+
+  namespace {
+    bool is_subword(std::string const& x, std::string const& y) {
+      return y.find(x) != std::string::npos;
+    }
+
+    void prefixes_string(std::unordered_map<std::string, size_t>& st,
+                         std::string const&                       x,
+                         size_t&                                  n) {
+      for (auto it = x.cbegin() + 1; it < x.cend(); ++it) {
+        auto w   = std::string(x.cbegin(), it);
+        auto wit = st.find(w);
+        if (wit == st.end()) {
+          st.emplace(w, n);
+          n++;
+        }
+      }
+    }
+
+  }  // namespace
+
   namespace fpsemigroup {
 
     //////////////////////////////////////////////////////////////////////////
@@ -70,6 +92,7 @@ namespace libsemigroups {
 
     KnuthBendix::KnuthBendix()
         : FpSemigroupInterface(),
+          _gilman_digraph(),
           _impl(detail::make_unique<KnuthBendixImpl>(this)) {}
 
     KnuthBendix::KnuthBendix(KnuthBendix const& kb) : KnuthBendix() {
@@ -79,7 +102,7 @@ namespace libsemigroups {
     KnuthBendix::~KnuthBendix() = default;
 
     //////////////////////////////////////////////////////////////////////////
-    // KnuthBendix - initialisers - public
+    // KnuthBendix - initialisers - private
     //////////////////////////////////////////////////////////////////////////
 
     void KnuthBendix::init_from(FroidurePinBase& S) {
@@ -113,13 +136,49 @@ namespace libsemigroups {
     // FpSemigroupInterface - non-pure virtual methods - public
     //////////////////////////////////////////////////////////////////////////
 
+    KnuthBendix::const_normal_form_iterator
+    KnuthBendix::cbegin_normal_forms(std::string const& lphbt,
+                                     size_t const       min,
+                                     size_t const       max) {
+      using state_type = NormalFormsIteratorTraits::state_type;
+      auto it          = const_normal_form_iterator(
+          state_type(lphbt, ""), gilman_digraph().cbegin_pislo(0, min, max));
+      if (min == 0 && !_impl->contains_empty_string()
+          && (!has_identity() || !identity().empty())) {
+        ++it;
+      }
+      return it;
+    }
+
     size_t KnuthBendix::size() {
       if (is_obviously_infinite()) {
         return POSITIVE_INFINITY;
       } else if (alphabet().empty()) {
         return 0;
       } else {
-        return froidure_pin()->size();
+        int const  modifier = (contains_empty_string() ? 0 : -1);
+        auto const out      = gilman_digraph().number_of_paths(0);
+        return (out == POSITIVE_INFINITY ? out : out + modifier);
+      }
+    }
+
+    // TODO move to the right section
+    // TODO noexcept?
+    bool KnuthBendix::contains_empty_string() const {
+      return _impl->contains_empty_string()
+             || (has_identity() && identity().empty());
+    }
+
+    // TODO move to the right section
+    uint64_t KnuthBendix::number_of_normal_forms(size_t const min,
+                                                 size_t const max) {
+      if (alphabet().empty()) {
+        return 0;
+      } else {
+        int const modifier = (contains_empty_string() ? 0 : -1);
+
+        auto const out = gilman_digraph().number_of_paths(0, min, max);
+        return (out == POSITIVE_INFINITY ? out : out + modifier);
       }
     }
 
@@ -131,6 +190,12 @@ namespace libsemigroups {
       validate_word(u);
       validate_word(v);
       return _impl->equal_to(u, v);
+    }
+
+    std::string KnuthBendix::normal_form(std::string const& w) {
+      validate_word(w);
+      run();
+      return rewrite(w);
     }
 
     //////////////////////////////////////////////////////////////////////////
@@ -180,6 +245,55 @@ namespace libsemigroups {
       report_why_we_stopped();
     }
 
+    std::vector<std::string> normal_forms(size_t min = 0,
+                                          size_t max = POSITIVE_INFINITY);
+
+    ActionDigraph<size_t> const& KnuthBendix::gilman_digraph() {
+      if (_gilman_digraph.nr_nodes() == 0) {
+        max_rules(POSITIVE_INFINITY);
+        run();
+        std::unordered_map<std::string, size_t> prefixes;
+        prefixes.emplace("", 0);
+        // FIXME
+        auto rules = active_rules();
+
+        size_t n = 1;
+        for (auto const& rule : rules) {
+          prefixes_string(prefixes, rule.first, n);
+        }
+
+        _gilman_digraph.add_nodes(prefixes.size());
+        _gilman_digraph.add_to_out_degree(alphabet().size());
+
+        for (auto& p : prefixes) {
+          for (size_t i = 0; i < alphabet().size(); ++i) {
+            auto s  = p.first + alphabet()[i];
+            auto it = prefixes.find(s);
+            if (it != prefixes.end()) {
+              _gilman_digraph.add_edge(p.second, it->second, i);
+            } else {
+              if (!std::any_of(
+                      rules.cbegin(),
+                      rules.cend(),
+                      [&s](std::pair<std::string, std::string> const& rule) {
+                        return is_subword(rule.first, s);
+                      })) {
+                while (!s.empty()) {
+                  s       = std::string(s.begin() + 1, s.end());
+                  auto it = prefixes.find(s);
+                  if (it != prefixes.end()) {
+                    _gilman_digraph.add_edge(p.second, it->second, i);
+                    break;
+                  }
+                }
+              }
+            }
+          }
+        }
+      }
+      return _gilman_digraph;
+    }
+
     //////////////////////////////////////////////////////////////////////////
     // FpSemigroupInterface - pure virtual methods - private
     //////////////////////////////////////////////////////////////////////////
@@ -200,13 +314,22 @@ namespace libsemigroups {
     }
 
     bool KnuthBendix::is_obviously_infinite_impl() {
-      if (alphabet().size() > nr_rules()) {
+      if (finished()) {
+        return !action_digraph_helper::is_acyclic(gilman_digraph());
+      } else if (alphabet().size() > nr_rules()) {
         return true;
       }
       detail::IsObviouslyInfinitePairs<char, std::string> ioi(
           alphabet().size());
       ioi.add_rules(cbegin_rules(), cend_rules());
       return ioi.result();
+    }
+
+    bool KnuthBendix::is_obviously_finite_impl() {
+      if (finished()) {
+        return action_digraph_helper::is_acyclic(gilman_digraph());
+      }
+      return false;
     }
 
     //////////////////////////////////////////////////////////////////////////
@@ -308,6 +431,10 @@ namespace libsemigroups {
       return const_contains(lhs, rhs) == tril::TRUE;
     }
 
+    bool KnuthBendix::is_quotient_obviously_infinite_impl() {
+      return _kb->is_obviously_infinite();
+    }
+
     ////////////////////////////////////////////////////////////////////////////
     // CongruenceInterface - pure virtual methods - private
     ////////////////////////////////////////////////////////////////////////////
@@ -340,7 +467,7 @@ namespace libsemigroups {
     }
 
     bool KnuthBendix::finished_impl() const {
-      return _kb->has_froidure_pin() && _kb->froidure_pin()->finished();
+      return _kb->finished();
     }
 
     ////////////////////////////////////////////////////////////////////////////
